@@ -1,59 +1,50 @@
-source 'http://rubygems.org'
+source 'https://rubygems.org'
 
-# gem 'rails', '3.0.10'
+gem 'rails', '3.2.13'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
 
-gem "sqlite3"
+group :development, :test do
+  gem 'sqlite3'
+end
 
-# Use unicorn as the web server
+
+# Gems used only for assets and not required
+# in production environments by default.
+group :assets do
+  gem 'sass-rails',   '~> 3.2.3'
+  gem 'coffee-rails', '~> 3.2.1'
+
+  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
+  # gem 'therubyracer', :platforms => :ruby
+
+  gem 'uglifier', '>= 1.0.3'
+end
+
+gem 'jquery-rails', '~> 2.0.0'
+
+# To use ActiveModel has_secure_password
+# gem 'bcrypt-ruby', '~> 3.0.0'
+
+# To use Jbuilder templates for JSON
+# gem 'jbuilder'
+
+# Use unicorn as the app server
 # gem 'unicorn'
 
 # Deploy with Capistrano
 # gem 'capistrano'
 
-# To use debugger (ruby-debug for Ruby 1.8.7+, ruby-debug19 for Ruby 1.9.2+)
-# gem 'ruby-debug'
-# gem 'ruby-debug19', :require => 'ruby-debug'
+# To use debugger
+# gem 'debugger'
 
-# Bundle the extra gems:
-# gem 'bj'
-# gem 'nokogiri'
-# gem 'sqlite3-ruby', :require => 'sqlite3'
-# gem 'aws-s3', :require => 'aws/s3'
+# Refinery CMS
+gem 'refinerycms', '~> 2.0.0', :git => 'git://github.com/refinery/refinerycms.git', :branch => '2-0-stable'
 
-# Bundle gems for the local environment. Make sure to
-# put test-only gems in this group so their generators
-# and rake tasks are available in development mode:
-# group :development, :test do
-#   gem 'webrat'
-# end
-
-# REFINERY CMS ================================================================
-# Anything you put in here will be overridden when the app gets updated.
-
-gem 'refinerycms',              '~> 1.0.8'
-
-group :development, :test do
-  # To use refinerycms-testing, uncomment it (if it's commented out) and run 'bundle install'
-  # Then, run 'rails generate refinerycms_testing' which will copy its support files.
-  # Finally, run 'rake' to run the tests.
-  # gem 'refinerycms-testing',    '~> 1.0.8'
-end
-
-# END REFINERY CMS ============================================================
-
-# USER DEFINED
-
-
-# Specify additional Refinery CMS Engines here (all optional):
-# gem 'refinerycms-inquiries',    '~> 1.0'
-# gem "refinerycms-news",         '~> 1.2'
-# gem 'refinerycms-blog',         '~> 1.6'
-# gem 'refinerycms-page-images',  '~> 1.0'
-
-# Add i18n support (optional, you can remove this if you really want to).
-gem 'refinerycms-i18n',         '~> 1.0.0'
-
-# END USER DEFINED
+# Specify additional Refinery CMS Extensions here (all optional):
+gem 'refinerycms-i18n', '~> 2.0.0'
+#  gem 'refinerycms-blog', '~> 2.0.0'
+#  gem 'refinerycms-inquiries', '~> 2.0.0'
+#  gem 'refinerycms-search', '~> 2.0.0'
+#  gem 'refinerycms-page-images', '~> 2.0.0'
