@@ -3,10 +3,8 @@
 #
 # Examples:
 #
-#   cities = City.create([{ :name => 'Chicago' }, { :name => 'Copenhagen' }])
-#   Mayor.create(:name => 'Daley', :city => cities.first)
-# Refinery seeds
-Dir[Rails.root.join('db', 'seeds', '*.rb').to_s].each do |file|
-  puts "Loading db/seeds/#{file.split(File::SEPARATOR).last}"
-  load(file)
-end
+#   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
+#   Mayor.create(name: 'Emanuel', city: cities.first)
+
+# Added by Refinery CMS Pages extension
+Refinery::Pages::Engine.load_seed
