@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Compass::Application.config.secret_token = 'e66f93c412f12e29da2323dcc720ff0e10c17189f0a70704f8dd9a05622f04bb9b8fe55d4baa29709cbe7dc8afe34657f7a097853074e09f200530a9a6e37d64'
+Compassremodeling::Application.config.secret_token = 'e98da16fcf2d561874574b389099874bb3622fc877d8ee47820606163afae9d985f789675f0e9a261877d85204978b95bf58a9d815f1b6715a50170c7e504071'
